% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_fellow}
\alias{check_fellow}
\title{Check Fellow}
\usage{
check_fellow(rota, fellow = "dermot")
}
\arguments{
\item{rota}{a rota list returned from \code{fit_rota()}}

\item{fellow}{a string of the fellow to check}
}
\value{

}
\description{
Check Fellow
}
