% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{encode_preferences}
\alias{encode_preferences}
\title{Title}
\usage{
encode_preferences(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
There is a potential target solution, but it doesn't strictly adhere to the
constraint functions. We can do this with GAs by defining a penalized fitness
function. So can we now use this method to optimise the rota. Let's start
with the simple case of 2 doctors, each to fill 2 days, with alternate day
preferences. I.e. the solution is already in their preferences.
}
