% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{fit_rota}
\alias{fit_rota}
\title{Fit Rota}
\usage{
fit_rota(
  file_path = NULL,
  fellows = c("matt", "muska", "zainab", "dermot", "justin", "paul"),
  nhs = c(0, 0, 0, 1, 0, 1),
  month = 7
)
}
\arguments{
\item{file_path}{}

\item{fellows}{}

\item{nhs}{}

\item{month}{}
}
\value{

}
\description{
Fit Rota
}
