% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{parse_preferences}
\alias{parse_preferences}
\title{Parse Preferences}
\usage{
parse_preferences(
  selections = NA_character_,
  requirement = NA_character_,
  fellows = c("matt", "muska", "zainab", "dermot", "ed", "justin", "paul"),
  nhs = c(0, 0, 0, 1, 1, 0, 1),
  month = 1
)
}
\arguments{
\item{selections}{character vector of the file path to the selection file}

\item{fellows}{a character vector containing the names of the fellows}

\item{nhs}{integer vector in the same order as "fellows" indicating if
certain fellows should be prioritised}

\item{month}{an integer vector of length 1 detailing the month of interest}

\item{requirements}{character vector of the file path to the requirements
file}
}
\value{

}
\description{
Parse Preferences
}
