% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{fit_rota}
\alias{fit_rota}
\title{Fit Rota}
\usage{
fit_rota(
  fellows = c("matt", "muska", "zainab", "dermot", "justin", "paul"),
  selections = NA_character_,
  requirement = NA_character_,
  nhs = c(0, 0, 0, 1, 0, 1),
  month = 7
)
}
\arguments{
\item{fellows}{a character vector containing the names of the fellows}

\item{selections}{character vector of the file path to the selection file}

\item{nhs}{integer vector in the same order as "fellows" indicating if
certain fellows should be prioritised}

\item{month}{an integer vector of length 1 detailing the month of interest}

\item{requirements}{character vector of the file path to the requirements
file}
}
\value{

}
\description{
Fit Rota
}
