% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness.R
\name{fitness_func}
\alias{fitness_func}
\title{Fitness Function}
\usage{
fitness_func(x, days, pref_strong, pref_weak, pref_no, pref_bigno)
}
\arguments{
\item{x}{}

\item{days}{}

\item{pref_strong}{}

\item{pref_weak}{}

\item{pref_no}{}

\item{pref_bigno}{}
}
\value{

}
\description{
Fitness Function
}
