% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{parse_preferences}
\alias{parse_preferences}
\title{Parse Preferences}
\usage{
parse_preferences(
  file_path = NULL,
  fellows = c("matt", "muska", "zainab", "dermot", "ed"),
  month = 1
)
}
\arguments{
\item{file_path}{a file path to the folder where prefernces are stored}

\item{fellows}{a character vector containing the names of the fellows}

\item{month}{an integer vector of length 1 detailing the month of interest}
}
\value{

}
\description{
Parse Preferences
}
